# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SindÂncora API",
        "description": "API REST do SindÂncora — SaaS de Gestão Condominial",
        "contact": {
            "email": "suporte@sindancora.com.br"
        },
        "license": {
            "name": "Proprietário"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Base"
        }
    ],
    "paths": {
...
        "/v1/app/gatehouse/visits": {
            "get": {
                "tags": [
                    "App"
                ],
                "summary": "Visitantes: presentes e histórico paginado (filtro: condominium_id)",
                "operationId": "appGatehouseVisits",
                "responses": {
                    "200": {
                        "description": "Presentes + histórico"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/app/gatehouse/validate-token": {
            "post": {
                "tags": [
                    "App"
                ],
                "summary": "Validar QR/token de autorização de visitante",
                "operationId": "appGatehouseValidate",
                "responses": {
                    "200": {
                        "description": "Resultado da validação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/app/gatehouse/check-in": {
            "post": {
                "tags": [
                    "App"
                ],
                "summary": "Registrar entrada: autorizado (authorization_id) ou avulso (dados do visitante)",
                "operationId": "appGatehouseCheckIn",
                "responses": {
                    "201": {
                        "description": "Entrada registrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/app/gatehouse/visits/{visit}/check-out": {
            "post": {
                "tags": [
                    "App"
                ],
                "summary": "Registrar saída do visitante",
                "operationId": "appGatehouseCheckOut",
                "parameters": [
                    {
                        "name": "visit",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Saída registrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/app/parcels": {
            "get": {
                "tags": [
                    "App"
                ],
                "summary": "Listar encomendas (filtros: condominium_id, status)",
                "operationId": "appParcelsIndex",
                "responses": {
                    "200": {
                        "description": "Lista paginada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            },
            "apiKey": {
                "type": "http",
                "description": "API Key do tenant no formato sk_live_... enviada como Bearer token.",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}